cleanup: move `logger.py` to `/observability/logging` (#541)
minor cleanup, non-functional change
