+++ inference_perf/loadgen/load_generator.py
@@ -76,7 +76,7 @@ from rich.progress import (
 )
 import signal
 
-from inference_perf.logger import get_console
+from inference_perf.observability.logging import get_console
 
 logger = logging.getLogger(__name__)
